Write some tests for Reciprocal Rank Fusion
